rancher project membership for groups

+++ src/Dev/Controllers/Kubernetes/Cluster/ProjectController.cs
@@ -3,7 +3,6 @@
 using DotnetKubernetesClient;
 using DotnetKubernetesClient.LabelSelectors;
 using Github.Internal;
-using k8s.Models;
 using KubeOps.Operator.Controller;
 using KubeOps.Operator.Controller.Results;
 using KubeOps.Operator.Rbac;
@@ -11,6 +10,8 @@ using v1.Components.Kubernetes;
 using v1.Core;
 using v1.Platform.Github;
 using v1.Platform.Rancher;
+using v1.Platform.Rancher.External;
+using Project = v1.Platform.Rancher.Project;
 using RancherProject = v1.Platform.Rancher.External.Project;
 
 [EntityRbac(typeof(Project), Verbs = RbacVerb.All)]
@@ -22,118 +23,97 @@ public class ProjectController : IResourceController<Project>
     public ProjectController(
         IKubernetesClient kubernetesClient,
         ILogger<ProjectController> logger
-        )
+    )
     {
         _kubernetesClient = kubernetesClient;
         _logger = logger;
     }
-    
+
     public async Task<ResourceControllerResult?> ReconcileAsync(Project? entity)
     {
         if (entity == null) return null;
-        
+
         var zoneName = entity.Metadata.NamespaceProperty;
-        
+
         var context = await _kubernetesClient.Get<TenancyContext>(TenancyContext.GetName(), zoneName);
         if (context == null) throw new Exception($"cannot find tenancy context for {zoneName}");
 
         var rancher = await _kubernetesClient.Get<Rancher>("rancher", context.Spec.OrganizationNamespace);
         if (rancher == null) throw new Exception($"cannot find Rancher in {context.Spec.OrganizationNamespace}");
-        
+
         var cluster = await _kubernetesClient.Get<Kubernetes>(entity.Spec.Kubernetes, zoneName);
         if (cluster == null) throw new Exception($"cannot find cluster {entity.Spec.Kubernetes}");
-        if (string.IsNullOrWhiteSpace(cluster.Status.ClusterId)) throw new Exception($"id has not been associated with cluster {entity.Spec.Kubernetes}");
+        if (string.IsNullOrWhiteSpace(cluster.Status.ClusterId))
+            throw new Exception($"id has not been associated with cluster {entity.Spec.Kubernetes}");
 
-        var label = $"{cluster.Status.ClusterId}/{entity.Metadata.Name}";
-        
+
+        var label = $"{cluster.Status.ClusterId}_{entity.Metadata.Name}";
+
         if (string.IsNullOrEmpty(entity.Status.Id))
         {
             var selector = new EqualsSelector(Project.ProjectLabel(), label);
-            IList<RancherProject> projects;
-            try
-            {
-                
-                projects = await _kubernetesClient.List<RancherProject>(cluster.Status.ClusterId, selector);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e);
-                throw;
-            }
-            
+            var projects = await _kubernetesClient.List<RancherProject>(cluster.Status.ClusterId, selector);
+
             var rancherProject = projects.FirstOrDefault() ?? await _kubernetesClient.Create(() =>
             {
-                
                 var p = new RancherProject();
                 RancherProject.Init(
-                    p, 
-                    entity.Metadata.Name, 
-                    rancher.Spec.TechnicalUser, 
+                    p,
+                    entity.Spec.Tenancy,
+                    rancher.Spec.TechnicalUser,
                     cluster.Status.ClusterId);
-
+                
                 p.Metadata.Labels ??= new Dictionary<string, string>();
                 p.Metadata.Labels.Add(Project.ProjectLabel(), label);
-                
+
                 return p;
-                
             }, null, cluster.Status.ClusterId);
 
             entity.Status.Id = rancherProject.Metadata.Name;
             await _kubernetesClient.UpdateStatus(entity);
         }
-        
+
         //setup the Github team to have access
-        
+
         //EXCEPTION TO A RULE
         //as im unsure how to query k8s for github groups/teams
-        var team = await _kubernetesClient.Get<Team>(entity.Spec.Tenancy, context.Spec.OrganizationNamespace);
-        if (team == null) throw new Exception($"cannot find github team for {entity.Spec.Tenancy}");
-        if (!team.Status.Id.HasValue)throw new Exception($"github team is not synced (yet) - {entity.Spec.Tenancy}");
+        var tenancy = entity.Spec.Tenancy;
+        var github = await _kubernetesClient.GetGithub(entity.Metadata.NamespaceProperty);
+        var name = github.IsGlobal(tenancy)
+            ? context.Spec.OrganizationNamespace // repo <- global team
+            : Team.GetTeamName(tenancy); // repo <- tenancy team(-guest)
 
-        var bindingName = $"{entity.Status.Id}-team-role";
-        var roleName = cluster.Status.IsProduction 
+
+        var team = await _kubernetesClient.Get<Team>(tenancy, name);
+        if (team == null) throw new Exception($"cannot find github team for {entity.Spec.Tenancy}");
+        if (!team.Status.Id.HasValue) throw new Exception($"github team is not synced (yet) - {entity.Spec.Tenancy}");
+        
+        var roleName = cluster.Status.IsProduction
             ? rancher.Spec.TenancyProductionMemberRole
             : rancher.Spec.TenancyMemberRole;
 
         //check, add or update....
-        var binding = await _kubernetesClient.Get<V1RoleBinding>(bindingName, cluster.Status.ClusterId);
+        var bindingSelector = new EqualsSelector(ProjectRoleTemplateBinding.RoleTenancy(), tenancy);
+        var bindings = await _kubernetesClient.List<ProjectRoleTemplateBinding>(entity.Status.Id, bindingSelector);
+        var binding = bindings.FirstOrDefault();
         if (binding == null)
         {
-            binding = new V1RoleBinding
+            await _kubernetesClient.Create(() =>  
             {
-                Metadata = new V1ObjectMeta
-                {
-                    Name = $"{entity.Status.Id}-team-role"
-                },
-                RoleRef = new V1RoleRef
-                {
-                    ApiGroup = "rbac.authorization.k8s.io",
-                    Kind = "role",
-                    Name = roleName
-                },
-                Subjects = new List<V1Subject>
-                {
-                    new()
-                    {
-                        ApiGroup = "rbac.authorization.k8s.io",
-                        Kind = "group",
-                        Name = $"github_org://{team.Status.Id}"
-                    }
-                }
-            };
+                var b = ProjectRoleTemplateBinding.InitGroup(
+                    rancher.Spec.TechnicalUser, 
+                    cluster.Status.ClusterId,
+                    entity.Status.Id, 
+                    tenancy,
+                    team.Status.Id.Value,
+                    roleName);
+                return b;
+            }, null, entity.Status.Id);
+        }
 
-            await _kubernetesClient.Create(binding);
-        }
-        
-        else if (binding.RoleRef.Name != roleName)
-        {
-            binding.RoleRef.Name = roleName;
-            await _kubernetesClient.Update(binding);
-        }
-        
         return null;
     }
-    
+
     public async Task DeletedAsync(Project? entity)
     {
         if (entity == null) return;
@@ -141,22 +121,24 @@ public class ProjectController : IResourceController<Project>
 
         var cluster = await _kubernetesClient.Get<Kubernetes>(entity.Spec.Kubernetes, zoneName);
         if (cluster == null) throw new Exception($"cannot find cluster {entity.Spec.Kubernetes}");
-        if (string.IsNullOrWhiteSpace(cluster.Status.ClusterId)) throw new Exception($"id has not been associated with cluster {entity.Spec.Kubernetes}");
+        if (string.IsNullOrWhiteSpace(cluster.Status.ClusterId))
+            throw new Exception($"id has not been associated with cluster {entity.Spec.Kubernetes}");
 
         var label = $"{cluster.Status.ClusterId}/{entity.Metadata.Name}";
-        
+
         var selector = new EqualsSelector(Project.ProjectLabel(), label);
 
         var rancherProjects = await _kubernetesClient
             .List<RancherProject>(
                 cluster.Status.ClusterId, //entity.Metadata.NamespaceProperty, 
                 selector);
-        
-        var rancherProject = rancherProjects.FirstOrDefault();
-        if (rancherProject == null) return;
 
-        await _kubernetesClient.Delete(rancherProject);
-        await _kubernetesClient.Delete<V1RoleBinding>($"{rancherProject.Metadata.Name}-team-role",
-            rancherProject.Metadata.NamespaceProperty);
+        var rancherProject = rancherProjects.FirstOrDefault();
+        if (rancherProject != null) await _kubernetesClient.Delete(rancherProject);;
+
+        var bindingSelector = new EqualsSelector(ProjectRoleTemplateBinding.RoleTenancy(), entity.Spec.Tenancy);
+        var bindings = await _kubernetesClient.List<ProjectRoleTemplateBinding>(entity.Status.Id, bindingSelector);
+        var binding = bindings.FirstOrDefault();
+        if (binding != null) await _kubernetesClient.Delete(binding);
     }
 }
